@@ -10,34 +10,63 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.0.3",
-    "@angular/common": "~12.0.3",
-    "@angular/compiler": "~12.0.3",
-    "@angular/core": "~12.0.3",
-    "@angular/forms": "~12.0.3",
-    "@angular/platform-browser": "~12.0.3",
-    "@angular/platform-browser-dynamic": "~12.0.3",
-    "@angular/router": "~12.0.3",
-    "apollo-angular": "^2.6.0",
+    "@angular/animations": "^13.0.0",
+    "@angular/cdk": "^13.0.0",
+    "@angular/common": "^13.0.0",
+    "@angular/compiler": "^13.0.0",
+    "@angular/core": "^13.0.0",
+    "@angular/flex-layout": "^12.0.0-beta.35",
+    "@angular/forms": "^13.0.0",
+    "@angular/material": "^13.0.0",
+    "@angular/platform-browser": "^13.0.0",
+    "@angular/platform-browser-dynamic": "^13.0.0",
+    "@angular/router": "^13.0.0",
+    "@apollo/client": "^3.4.16",
+    "@ckeditor/ckeditor5-angular": "^2.0.2",
+    "@ckeditor/ckeditor5-build-classic": "^29.2.0",
+    "@ecodev/fab-speed-dial": "^8.0.1",
+    "@fortawesome/angular-fontawesome": "^0.9.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.36",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
+    "@ng-bootstrap/ng-bootstrap": "^10.0.0",
+    "@turf/centroid": "^6.5.0",
+    "@turf/helpers": "^6.5.0",
+    "@turf/union": "^6.5.0",
+    "@types/d3": "^7.1.0",
+    "@types/file-saver": "^2.0.5",
+    "@types/ol": "^6.5.3",
+    "@types/proj4": "^2.5.2",
+    "bootstrap-css-only": "^4.4.1",
+    "d3": "^7.1.1",
+    "file-saver": "^2.0.5",
+    "graphql": "^15.7.2",
+    "ngx-color-picker": "^11.0.0",
+    "ngx-cookie-service": "^12.0.3",
+    "ol": "^6.9.0",
+    "proj4": "^2.7.5",
     "rxjs": "~6.6.0",
-    "tslib": "^2.1.0",
-    "zone.js": "~0.11.4",
-    "@apollo/client": "^3.0.0",
-    "graphql": "^15.0.0"
+    "save-svg-as-png": "^1.4.17",
+    "tinycolor2": "^1.4.2",
+    "tslib": "^2.3.1",
+    "uuid": "^8.3.2",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.0.3",
-    "@angular/cli": "~12.0.3",
-    "@angular/compiler-cli": "~12.0.3",
-    "@angular/localize": "^12.0.3",
+    "@angular-devkit/build-angular": "^13.2.3",
+    "@angular/cli": "^13.2.3",
+    "@angular/compiler-cli": "^13.0.0",
+    "@angular/localize": "^13.0.0",
     "@types/jasmine": "~3.6.0",
-    "@types/node": "^12.11.1",
+    "@types/node": "^12.20.36",
+    "@types/save-svg-as-png": "github:ksholla20/save-svg-as-png-typings",
+    "@types/tinycolor2": "^1.4.3",
+    "@types/uuid": "^8.3.4",
     "jasmine-core": "~3.7.0",
-    "karma": "~6.3.0",
+    "karma": "^6.3.16",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "typescript": "~4.2.3"
+    "karma-jasmine-html-reporter": "^1.7.0",
+    "typescript": "~4.4.4"
   }
 }
