@@ -3,7 +3,9 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false
+  production: false,
+  backend: 'http://localhost:8000',
+  apiPath: 'http://localhost:8000/api'
 };
 
 /*
