@@ -1,32 +1,43 @@
 from datentool.settings import *
 
+DEBUG = True
+
 INSTALLED_APPS.extend([
     'corsheaders'
 ])
 
-MIDDLEWARE.extend([
+CORS_ORIGIN_ALLOW_ALL = True
+CORS_ALLOW_CREDENTIALS = True
+ALLOWED_HOSTS = ['*']
+#ALLOWED_HOSTS.extend([
+    #'localhost',
+    #'127.0.0.1',
+    #'0.0.0.0'
+#])
+
+# cors midleware has to be loaded first
+MIDDLEWARE = [
     'corsheaders.middleware.CorsMiddleware',
     'django.middleware.common.CommonMiddleware'
+] + MIDDLEWARE
+
+# allow access to Rest API with session in development only
+# (in production only auth. via tokens is allowed)
+REST_FRAMEWORK['DEFAULT_AUTHENTICATION_CLASSES'].extend([
+    'rest_framework.authentication.SessionAuthentication',
 ])
-
-CORS_ORIGIN_ALLOW_ALL = True
-
-DEBUG = True
-
-# GDAL
-if os.name == 'nt':
-    # alternate install via conda
-    os.environ['GDAL_DATA'] = os.path.join(sys.prefix, 'Library',
-                                           'share', 'gdal')
-    os.environ['PATH'] = ';'.join([os.environ['PATH'],
-                                   os.path.join(sys.prefix, 'Library', 'bin')])
 
 DATABASES['default']['OPTIONS']['sslmode'] = 'prefer'
 
-# SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = 'django-insecure-mzejv_pa9tbj7$5$q%ju0ko*)vrouq3_+0&q)y@phi!fevpntp'
+# default secret keys, for dev only!
+SECRET_KEY = os.environ.get(
+    'SECRET_KEY',
+    'django-insecure-mzejv_pa9tbj7$5$q%ju0ko*)vrouq3_+0&q)y@phi!fevpntp'
+)
+# dummy encryption key - don't use this one in production
+ENCRYPT_KEY = os.environ.get('ENCRYPT_KEY',
+                             '85Ao9Yoarq05G2p_QsD_zDg5FTdOc3OaA2SndOBP55s=')
 
-ALLOWED_HOSTS.extend([
-    'localhost',
-    '127.0.0.1'
-])
+# set to False to run django_q tasks synchronous in dev mode without
+# needing running seperate q-cluster
+USE_DJANGO_Q = True
