@@ -1,3 +1,10 @@
-from django.contrib import admin
-
-# Register your models here.
+import datentool_backend.area.admin
+import datentool_backend.user.admin
+import datentool_backend.modes.admin
+import datentool_backend.indicators.admin
+import datentool_backend.logging.admin
+import datentool_backend.infrastructure.admin
+import datentool_backend.population.admin
+import datentool_backend.site.admin
+import datentool_backend.demand.admin
+import datentool_backend.places.admin
