@@ -1,3 +1,9 @@
-from django.test import TestCase
-
-# Create your tests here.
+from .user.tests import *
+from .area.tests import *
+from .modes.tests import *
+from .population.tests import *
+from .infrastructure.tests import *
+from .indicators.tests import *
+from .demand.tests import *
+from .places.tests import *
+from .site.tests import *
