@@ -1,3 +1,10 @@
-from django.shortcuts import render
-
-# Create your views here.
+from .area.views import *
+from .user.views import *
+from .population.views import *
+from .modes.views import *
+from .demand.views import *
+from .infrastructure.views import *
+from .site.views import *
+from .indicators.views import *
+from .places.views import *
+from .logging.views import *
